Gate to arm only devices and fix nitpicks.
There could potentially be issues with the native library being
loaded on things not compatible with `armeabi-v7a`. The library
should not be loaded and the code won't be called if this is the case.
This should close issue #12.

Due to an API I'm using to do the above gating, the minimum is getting
bumped from `android-16` to `android-21` now.

Also fixed some implicite declarations and bad return types in the c
code.